win fixed, remove several warnings

+++ VirtualRobot/Model/Obstacle.cpp
@@ -292,7 +292,7 @@ namespace VirtualRobot
     }
 
 
-    void Obstacle::setMass(int mass)
+    void Obstacle::setMass(float mass)
     {
         WriteLockPtr l = getWriteLock();
         auto links = getLinks();
